Added GB/GBC shared cartridge loading
